parser/tests/tree_construction/basic.dat

#data
<b><table><td></b><i></table>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,18): unexpected-end-tag
(1,29): unexpected-cell-end-tag
(1,30): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>
|       "X"
